+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Merged `sess` and `stats` in to single CLI command
 - Changed `sess --stats` to `stats list`
 - Ability to filter sessions by client IDs
+- New option `--with-key` with `clients list`
+- Changed `--logger` to `logger-id` in `rconfig` for consistency
 
 ## [1.4.0]
 ### Updates